move anchors of siamrpnpp to utils
